requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==41.0.5
cssselect2==0.7.0
Django==4.2.7
django-cors-headers==4.3.0
django-crispy-forms==2.1
djangorestframework==3.14.0
docopt==0.6.2
html5lib==1.1
idna==3.4
lxml==4.9.3
oscrypto==1.3.0
Pillow==10.1.0
pipreqs==0.4.13
psycopg2-binary==2.9.9
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyMySQL==1.1.0
pypdf==3.17.0
pypng==0.20220715.0
pyscopg2==66.0.2
python-bidi==0.4.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.0.7
webencodings==0.5.1
xhtml2pdf==0.2.11
yarg==0.1.9
